Split samplers implementation across different files (#259)
